Refactor: Split AuthContext into smaller files
Refactor AuthContext to improve code organization and readability by splitting it into separate files for different authentication methods.
